// src/models/apflora/Ekfrequenz.ts
export type EkfrequenzId = string & { __brand: 'apflora.ekfrequenz' };

/** Represents the table apflora.ekfrequenz */
export default interface Ekfrequenz {
  /** Primärschlüssel */
  id: EkfrequenzId;

  /** Zugehöriger Aktionsplan. Fremdschlüssel aus der Tabelle "ap" */
  ap_id: ApId;

  /** Beschreibt, in welchen Fällen diese Frequenz angewandt wird. Wahrscheinliche Werte: autochthone Population, angepflanzte Population, angesäte Population, Spezialfall */
  anwendungsfall: string | null;

  /** Definierend für die eqfrequenz */
  code: string | null;

  /** Definiert, in welchen Jahren eine Kontrolle üblicherweise stattfinden soll. Bei Anpflanzungen sind das Jahre ab der letzten Anpflanzung. Bei autochthonen Populationen? */
  kontrolljahre: number[] | null;

  /** Ab dieser Anzahl Individuen wird diese Frequenz bei autochthonen Populationen (normalerweise) gewählt. Bei Anpflanzungen nicht relevant */
  anzahl_min: number | null;

  /** Bis zu dieser Anzahl Individuen wird diese Frequenz bei autochthonen Populationen (normalerweise) gewählt. Bei Anpflanzungen nicht relevant */
  anzahl_max: number | null;

  bemerkungen: string | null;

  /** Damit EK-Zähleinheiten untereinander sortiert werden können */
  sort: number | null;

  /** Wer hat den Datensatz zuletzt geändert? */
  changed_by: string | null;

  /** Fremdschlüssel aus Tabelle ek_abrechnungstyp_werte. Bestimmt, wie Kontrollen abgerechnet werden sollen */
  ek_abrechnungstyp: string | null;

  /** Ob diese Frequenz für EK oder EKF anwendbar ist */
  ektyp: public_EkType | null;

  /** Referenzjahr für die Kontrolljahre */
  kontrolljahre_ab: public_EkKontrolljahreAb | null;

  created_at: Date;

  updated_at: Date;
}